refactor: movendo arquivos para nova estrutura de pastas
