+++ api/web/static/scripts.js
